server: require hook hotfix  (#55146)
This PR fixes a bug in the require hook implementation where the variable used might not exist depending on the environment. In order to fix that, we can use the variable used to define the version of React used instead as it is already present and a proxy of the same information.




Co-authored-by: JJ Kasper <22380829+ijjk@users.noreply.github.com>