Ensure `ImageResponse` extends `Response` (#55187)
### Why?

The `ImageResponse` class constructor returns a `Response` instance, but doesn't extend `Response`, so it doesn't pass the type check introduced in #51394.

### How?

Make the `ImageResponse` class a subclass of `Response`.

### Notes

I can't find any tests to check for the types so didn't add any tests here; but shouldn't there be one?