Router worker: send SIGINT when the parent process terminates
Previously, `spawnSync` would block the parent process from handling signals. This refactors the spawn code to be asynchronous, sending SIGINT when the parent process is exiting or is receives an exit signal.

The next-dev command relied on the `spawnSync` blocking and preventing it from running that way. Instead, this now centralizes the check for whether there should be a router worker, and the next-dev command returns early if that's the case.

Test Plan: Ran Turbopack's benchmarks, which previously resulted in attempted file reads by the worker process when the main process quit and sources were removed.